+++ src-tauri/tauri.conf.json
@@ -69,7 +69,7 @@
         "fullscreen": false,
         "resizable": true,
         "title": "Noteup",
-        "url": "/app",
+        "url": "/notebook",
         "width": 1200,
         "height": 1200,
         "minWidth": 1000,
