+++ src/pages/Notebook/NotePreviewer/style.ts
