+++ src/pages/Notebook/Notebook.tsx
@@ -1,7 +1,7 @@
 import { useState } from "react";
 
-import { NoteContainer } from "@/Notebook/NoteContainer";
-import { SettingsModal } from "@/Notebook/SettingsModal";
+import { NoteContainer } from "@/pages/Notebook/NoteContainer";
+import { SettingsModal } from "@/pages/Notebook/SettingsModal";
 
 import { KeyboardShortcuts } from "@/components/KeyboardShortcuts";
 import { MobileNav } from "@/components/MobileNav";
@@ -11,9 +11,7 @@ import { SplitPane } from "@/components/SplitPanel";
 
 import { Container } from "@/styles/layout";
 
-import Navigation from '@/components/Layout/Navigation'
-
-export const AppContainer = () => {
+export const Notebook = () => {
   const [showSettings, setShowSettings] = useState(false);
 
   return (
