+++ src/pages/Notebook/SplitScreenEditor.tsx
